Add ability to pull from storage and display tasks
A user should be able to see the tasks that have been created and stored
in storage. The display reflects what is in storage by pulling from
storage any time a new task is created. The new task is sent to storage
and the storage is read and displayed on the page.