Use date-fns to change date formats
The formats for the dates being used to create tasks from the input form
was creating inconsistent behavior where the task due-date was one day
behind after creating the task. The format of the date from the form
is changed form hyphens to forward slashes and date-fns is used to
remove time from the date times in a more consistent way.