Fix cancel button functionality in edit mode
The cancel button was not resubmitting the populated data to recreate
the task. The added code allows for a new task to be created with old
data. It also creates a pipeline for the task id and create date to be
transferred from the old task to the edited task as that data needs to
be retained to remake the task.